Move global pipeline loading to pipeline module
